refactor: moved some helper functions to shared folder
