+++ forward_engineering/alterScript/alterScriptHelpers/columnHelpers/defaultValueHelper.js
@@ -1,6 +1,7 @@
 const _ = require('lodash');
 const { AlterScriptDto } = require('../../types/AlterScriptDto');
-const { generateFullEntityName, prepareName } = require('../../../utils/general');
+const { generateFullEntityName } = require('../../../utils/general');
+const { prepareName } = require('../../../../shared/general');
 
 /**
  * @return {({ collection, dbVersion }: { collection: Object, dbVersion: string }) => Array<AlterScriptDto>}
