+++ shared/constants.js
@@ -11,7 +11,15 @@ const COMMAND_EXECUTION_STATUS = {
 const REQUEST_TIMEOUT_MESSAGE =
 	'Request timeout exceeded, please try again or increase query request timeout in Tools > Options > Reverse-Engineering';
 
+/** How many columns to pull nullable/index metadata per Python command (avoids “Results too large”). */
+const FIELD_METADATA_COLUMN_BATCH_SIZE = 12;
+
+/** Truncate long strings in Spark field metadata when serializing indexes (per batch). */
+const FIELD_METADATA_STRING_MAX = 4000;
+
 module.exports = {
 	COMMAND_EXECUTION_STATUS,
 	REQUEST_TIMEOUT_MESSAGE,
+	FIELD_METADATA_COLUMN_BATCH_SIZE,
+	FIELD_METADATA_STRING_MAX,
 };
